refactor(snowflake): 重构雪花ID生成器配置与模型ID类型
- 将雪花节点初始化从硬编码改为从配置文件读取
- 统一各服务中雪花ID生成器的初始化调用
- 将模型ID类型从uint改为int64以兼容雪花ID
- 移除各服务中独立的雪花节点实例，改用全局生成器
- 更新相关测试代码以适配新的初始化方式

+++ pkg/common/repo.go
@@ -24,7 +24,7 @@ func (r *BaseRepo[T]) Delete(entity *T) error {
 	return r.DB.Delete(entity).Error
 }
 
-func (r *BaseRepo[T]) FindByID(id uint) (*T, error) {
+func (r *BaseRepo[T]) FindByID(id int64) (*T, error) {
 	var entity T
 	err := r.DB.First(&entity, id).Error
 	return &entity, err
